remove in-browser devtools, redux-thunk
