+++ examples/real-world/package.json
@@ -4,9 +4,6 @@
   "private": true,
   "devDependencies": {
     "react-scripts": "^4.0.0",
-    "redux-devtools": "^3.7.0",
-    "redux-devtools-dock-monitor": "^1.2.0",
-    "redux-devtools-log-monitor": "^2.1.0",
     "redux-logger": "^3.0.6"
   },
   "dependencies": {
@@ -18,8 +15,7 @@
     "react-dom": "^17.0.1",
     "react-redux": "^7.2.1",
     "react-router-dom": "^5.2.0",
-    "redux": "^4.0.5",
-    "redux-thunk": "^2.3.0"
+    "redux": "^4.0.5"
   },
   "scripts": {
     "start": "react-scripts start",
