+++ examples/real-world/src/actions/index.js
@@ -15,9 +15,9 @@ export const loadUser = createAsyncThunk(
     }
   },
   {
-    // If the user is already cached, this thunk will not dispatch anything.
     condition: ({ login, requiredFields }, { getState }) => {
       const user = getState().entities.users[login]
+      // If the user is already cached, nothing will be requested or dispatched
       const shouldProceed = !(
         user && requiredFields.every(key => user.hasOwnProperty(key))
       )
@@ -40,9 +40,9 @@ export const loadRepo = createAsyncThunk(
     }
   },
   {
-    // If the user is repo cached, this thunk will not dispatch anything.
     condition: ({ fullName, requiredFields }, { getState }) => {
       const repo = getState().entities.repos[fullName]
+      // If the repo is cached, nothing will be requested or dispatched
       const shouldProceed = !(
         repo && requiredFields.every(key => repo.hasOwnProperty(key))
       )
@@ -51,8 +51,9 @@ export const loadRepo = createAsyncThunk(
   }
 )
 
-// Fetches a page of starred repos by a particular user.
-// Bails out if page is cached and user didn't specifically request next page.
+/**
+ * Fetches a page of starred repos by a particular user.
+ */
 export const loadStarred = createAsyncThunk(
   'loadStarred',
   async ({ login }, { rejectWithValue, getState }) => {
@@ -66,9 +67,9 @@ export const loadStarred = createAsyncThunk(
     }
   },
   {
-    // If the user is repo cached, this thunk will not dispatch anything.
     condition: ({ login, nextPage }, { getState }) => {
       const { pageCount = 0 } = getState().pagination.starredByUser[login] || {}
+      // Bails out if page is cached and user didn't specifically request next page.
       const shouldProceed = !(pageCount > 0 && !nextPage)
       return shouldProceed
     }
@@ -77,8 +78,6 @@ export const loadStarred = createAsyncThunk(
 
 /**
  * Fetches a page of stargazers for a particular repo.
- * Bails out if page is cached and user didn't specifically request next page.
- * Relies on Redux Thunk middleware.
  */
 export const loadStargazers = createAsyncThunk(
   'loadStargazers',
@@ -93,10 +92,10 @@ export const loadStargazers = createAsyncThunk(
     }
   },
   {
-    // If the user is repo cached, this thunk will not dispatch anything.
     condition: ({ nextPage, fullName }, { getState }) => {
       const { pageCount = 0 } =
         getState().pagination.stargazersByRepo[fullName] || {}
+      // Bails out if page is cached and user didn't specifically request next page.
       const shouldProceed = !(pageCount > 0 && !nextPage)
       return shouldProceed
     }
