--- examples/real-world/src/store/hasDevtoolsExtension.js
@@ -1,3 +1,0 @@
-const hasDevtoolsExtension =
-  typeof window.__REDUX_DEVTOOLS_EXTENSION__ !== 'undefined'
-export default hasDevtoolsExtension
